master: no message

+++ bash-util.sh
@@ -413,11 +413,13 @@ function envsToSimpleEnvs()
 
     BUILD_IMAGE_NAME=$(sed "s/.dockerfile//g" <<< "${BUILD_IMAGE_NAME}")
 
-    FILE_SRC=/tmp/$(basename ${FILE})
-    rm -rf ${FILE_SRC}
-    cp -r ${FILE} ${FILE_SRC}
-    sort ${FILE_SRC}>${FILE}
-    rm -rf ${FILE_SRC}
+    if [[ -f ${FILE} ]]; then
+      FILE_SRC=/tmp/$(basename ${FILE})
+      rm -rf ${FILE_SRC}
+      cp -r ${FILE} ${FILE_SRC}
+      sort ${FILE_SRC}>${FILE}
+      rm -rf ${FILE_SRC}
+    fi
 
     logSuccess ${idt}
     logFinished ${idt} "envsToSimpleEnvs"
